+++ code_generation/templates/power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/attribute_classes.hpp.jinja
@@ -11,9 +11,9 @@
 
 #include "meta_data.hpp"
 
-#include "../enum.hpp"
-#include "../power_grid_model.hpp"
-#include "../three_phase_tensor.hpp"
+#include "../common/enum.hpp"
+#include "../common/power_grid_model.hpp"
+#include "../common/three_phase_tensor.hpp"
 
 namespace power_grid_model {
