+++ code_generation/templates/power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/meta_gen/attribute_classes.hpp.jinja
@@ -11,9 +11,9 @@
 
 #include "gen_getters.hpp" // NOLINT
 
-#include "../../enum.hpp"               // NOLINT
-#include "../../power_grid_model.hpp"   // NOLINT
-#include "../../three_phase_tensor.hpp" // NOLINT
+#include "../../common/enum.hpp"               // NOLINT
+#include "../../common/power_grid_model.hpp"   // NOLINT
+#include "../../common/three_phase_tensor.hpp" // NOLINT
 #include "../meta_data.hpp"             // NOLINT
 #include "../{{ name }}.hpp" // NOLINT
