+++ power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/meta_data.hpp
@@ -8,8 +8,8 @@
 
 #include "../common/enum.hpp"
 #include "../common/power_grid_model.hpp"
+#include "../common/three_phase_tensor.hpp"
 #include "../exception.hpp"
-#include "../three_phase_tensor.hpp"
 
 #include <bit>
 #include <span>
