+++ power_grid_model_c/power_grid_model/include/power_grid_model/auxiliary/update.hpp
@@ -13,7 +13,7 @@
 
 #include "../common/enum.hpp"
 #include "../common/power_grid_model.hpp"
-#include "../three_phase_tensor.hpp"
+#include "../common/three_phase_tensor.hpp"
 
 namespace power_grid_model {
