+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/load_gen.hpp
@@ -14,8 +14,8 @@
 #include "../auxiliary/update.hpp"
 #include "../calculation_parameters.hpp"
 #include "../common/power_grid_model.hpp"
+#include "../common/three_phase_tensor.hpp"
 #include "../exception.hpp"
-#include "../three_phase_tensor.hpp"
 
 namespace power_grid_model {
