+++ power_grid_model_c/power_grid_model/include/power_grid_model/component/voltage_sensor.hpp
@@ -10,8 +10,8 @@
 
 #include "../calculation_parameters.hpp"
 #include "../common/power_grid_model.hpp"
+#include "../common/three_phase_tensor.hpp"
 #include "../exception.hpp"
-#include "../three_phase_tensor.hpp"
 
 #include <limits>
