+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/block_matrix.hpp
@@ -10,7 +10,7 @@
 // the name of block matrix getter can be overwritten by sub matrix
 
 #include "../common/power_grid_model.hpp"
-#include "../three_phase_tensor.hpp"
+#include "../common/three_phase_tensor.hpp"
 
 namespace power_grid_model::math_solver {
