+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/iterative_current_pf_solver.hpp
@@ -62,8 +62,8 @@ Nomenclature:
 
 #include "../calculation_parameters.hpp"
 #include "../common/power_grid_model.hpp"
+#include "../common/three_phase_tensor.hpp"
 #include "../exception.hpp"
-#include "../three_phase_tensor.hpp"
 #include "../timer.hpp"
 
 namespace power_grid_model::math_solver {
