+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/newton_raphson_pf_solver.hpp
@@ -153,8 +153,8 @@ J.L -= -dQ_cal_m/dV
 
 #include "../calculation_parameters.hpp"
 #include "../common/power_grid_model.hpp"
+#include "../common/three_phase_tensor.hpp"
 #include "../exception.hpp"
-#include "../three_phase_tensor.hpp"
 #include "../timer.hpp"
 
 namespace power_grid_model::math_solver {
