+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/newton_raphson_se_solver.hpp
@@ -14,7 +14,7 @@
 #include "y_bus.hpp"
 
 #include "../calculation_parameters.hpp"
-#include "../three_phase_tensor.hpp"
+#include "../common/three_phase_tensor.hpp"
 
 namespace power_grid_model::math_solver {
