+++ power_grid_model_c/power_grid_model/include/power_grid_model/math_solver/y_bus.hpp
@@ -8,7 +8,7 @@
 
 #include "../calculation_parameters.hpp"
 #include "../common/power_grid_model.hpp"
-#include "../three_phase_tensor.hpp"
+#include "../common/three_phase_tensor.hpp"
 
 #include <algorithm>
 #include <ranges>
